# Reduce compile time for generic matmatmul