--- a/stdlib/LinearAlgebra/src/adjtrans.jl
+++ b/stdlib/LinearAlgebra/src/adjtrans.jl
@@ -283,6 +283,9 @@ adjoint(A::Transpose{<:Real}) = A.parent
 transpose(A::Adjoint{<:Real}) = A.parent
 adjoint(A::Transpose{<:Any,<:Adjoint}) = transpose(A.parent.parent)
 transpose(A::Adjoint{<:Any,<:Transpose}) = adjoint(A.parent.parent)
+# disambiguation
+adjoint(A::Transpose{<:Real,<:Adjoint}) = transpose(A.parent.parent)
+transpose(A::Adjoint{<:Real,<:Transpose}) = adjoint(A.parent.parent)
 
 # printing
 function Base.showarg(io::IO, v::Adjoint, toplevel)