@oscardssmith 2023-11-05
this looks great! are all the constprop afterwards still needed now that the characters are introduced lower down?